@@ -1,14 +1,14 @@
 module bitbucket-cascade-merge
 
-// +heroku goVersion go1.15
-go 1.15
+go 1.23.0
 
 require (
-	github.com/gin-gonic/gin v0.0.0-20150626140855-4cc2de6207f4
-	github.com/heroku/x v0.0.0-20171004170240-705849e307dd
-	github.com/ktrysmt/go-bitbucket v0.6.5
-	github.com/manucorporat/sse v0.0.0-20150604091100-c142f0f1baea // indirect
-	github.com/stretchr/testify v1.3.0 // indirect
-	golang.org/x/mod v0.3.0
-	gopkg.in/bluesuncorp/validator.v5 v5.9.1 // indirect
+	github.com/ktrysmt/go-bitbucket v0.9.81
+	golang.org/x/mod v0.22.0
+)
+
+require (
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	golang.org/x/net v0.32.0 // indirect
+	golang.org/x/oauth2 v0.24.0 // indirect
 )
